Add files via upload
Designed a rock, paper, scissors game in which you play against a computer with different colors at the output . You have the option to restart a game directly without having to run the program again. The points are also counted if you win. If you lose, the previous points are displayed and the score is set to 0. With "Draw", the previous score remains unchanged.